Initial commit
This is the first commit of the barcode_finder application. It is a
video4linux2 image grabber application which runs a FFT on each scan
line in the grabbed image in hopes of locating a 1D barcode (TODO).

main.c started out as a copy of v4l2grab, from here
https://github.com/twam/v4l2grab