Make MJPEG code path more robust
Check that the frame we get starts with 0xff 0xd8. Also, the JFIF header
is not needed, so remove it.

To sum up, to make a valid JPEG file from a webcam MPEG frame (so that
libjpeg can decompress it), simply insert the fixed huffman table right
after the start-of-image (SOI) marker (0xff 0xd8).